auth_app/requirements.txt
annotated-types
anyio
bcrypt
cffi
click
colorama
cryptography
dnspython
ecdsa
email_validator
fastapi
greenlet
h11
idna
loguru
motor
passlib
pyasn1
pycparser
pydantic
pydantic_core
pymongo
python-dotenv
python-jose
python-multipart
pytz
rsa
six
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
tzdata
uvicorn
win32_setctime
